Separated tests to individual states
