+++ src/UnitTests/PubnubApi.Tests/EventEngine/UnsubscribedStateTransition.cs
@@ -1,0 +1,61 @@
+﻿using NUnit.Framework;
+using PubnubApi.EventEngine.Core;
+using PubnubApi.EventEngine.Subscribe.Common;
+using PubnubApi.EventEngine.Subscribe.Context;
+using PubnubApi.EventEngine.Subscribe.Events;
+using PubnubApi.EventEngine.Subscribe.States;
+using System.Linq;
+
+namespace PubnubApi.Tests.EventEngine
+{
+    internal class UnsubscribedStateTransition
+    {
+        [Test]
+        public void TestUnsubscribedStateTransitionWithSubscriptionChangedEvent()
+        {
+            //Arrange
+            State unsubscribeState = new UnsubscribedState() { Channels = new string[] { "ch1", "ch2" }, ReconnectionConfiguration = new ReconnectionConfiguration(PNReconnectionPolicy.LINEAR, 50) };   
+            State handshakingState = new HandshakingState();
+            //Act
+            TransitionResult result = unsubscribeState.Transition(new SubscriptionChangedEvent()
+            {
+                Channels = new string[] { "ch1", "ch2" },
+                ChannelGroups = new string[] { "cg1", "cg2" }
+            });
+            //Assert
+            Assert.IsTrue(result.State.GetType().Equals(handshakingState.GetType()));
+            Assert.AreEqual("ch1", ((HandshakingState)(result.State)).Channels.ElementAt(0));
+            Assert.AreEqual("ch2", ((HandshakingState)(result.State)).Channels.ElementAt(1));
+            Assert.AreEqual("cg1", ((HandshakingState)(result.State)).ChannelGroups.ElementAt(0));
+            Assert.AreEqual("cg2", ((HandshakingState)(result.State)).ChannelGroups.ElementAt(1));
+            Assert.AreEqual(PNReconnectionPolicy.LINEAR, ((HandshakingState)(result.State)).ReconnectionConfiguration.ReconnectionPolicy);
+            Assert.AreEqual(50, ((HandshakingState)(result.State)).ReconnectionConfiguration.MaximumReconnectionRetries);
+        }
+        
+        [Test]
+        public void TestUnsubscribedStateTransitionWithSubscriptionRestoreEvent()
+        {
+            //Arrange
+            State unsubscribeState = new UnsubscribedState() { Channels = new string[] { "ch1", "ch2" }, ChannelGroups = new string[] { "cg1", "cg2" }, ReconnectionConfiguration = new ReconnectionConfiguration(PNReconnectionPolicy.LINEAR, 50) };
+            State receivingState = new ReceivingState();
+            //Act
+            TransitionResult result = unsubscribeState.Transition(new SubscriptionRestoredEvent()
+            {
+                Channels = new string[] { "ch1", "ch2" },
+                ChannelGroups = new string[] { "cg1", "cg2" },
+                Cursor = new SubscriptionCursor() { Region = 1, Timetoken = 1234567890 }
+            });
+            //Assert
+            Assert.IsTrue(result.State.GetType().Equals(receivingState.GetType()));
+            Assert.AreEqual("ch1", ((ReceivingState)(result.State)).Channels.ElementAt(0));
+            Assert.AreEqual("ch2", ((ReceivingState)(result.State)).Channels.ElementAt(1));
+            Assert.AreEqual("cg1", ((ReceivingState)(result.State)).ChannelGroups.ElementAt(0));
+            Assert.AreEqual("cg2", ((ReceivingState)(result.State)).ChannelGroups.ElementAt(1));
+            Assert.AreEqual(1, ((ReceivingState)(result.State)).Cursor.Region);
+            Assert.AreEqual(1234567890, ((ReceivingState)(result.State)).Cursor.Timetoken);
+            Assert.AreEqual(PNReconnectionPolicy.LINEAR, ((ReceivingState)(result.State)).ReconnectionConfiguration.ReconnectionPolicy);
+            Assert.AreEqual(50, ((ReceivingState)(result.State)).ReconnectionConfiguration.MaximumReconnectionRetries);
+        }
+
+    }
+}
