+++ src/UnitTests/PubnubApiPCL.Tests/PubnubApiPCL.Tests.csproj
@@ -51,7 +51,15 @@
   <ItemGroup>
     <Compile Include="..\PubnubApi.Tests\CleanupGrant.cs" Link="CleanupGrant.cs" />
     <Compile Include="..\PubnubApi.Tests\EncryptionTests.cs" Link="EncryptionTests.cs" />
-    <Compile Include="..\PubnubApi.Tests\EventEngineTests.cs" Link="EventEngineTests.cs" />
+    <Compile Include="..\PubnubApi.Tests\EventEngine\HandshakeFailedStateTransition.cs" Link="EventEngine\HandshakeFailedStateTransition.cs" />
+    <Compile Include="..\PubnubApi.Tests\EventEngine\HandshakeReconnectingStateTransition.cs" Link="EventEngine\HandshakeReconnectingStateTransition.cs" />
+    <Compile Include="..\PubnubApi.Tests\EventEngine\HandshakeStoppedStateTransition.cs" Link="EventEngine\HandshakeStoppedStateTransition.cs" />
+    <Compile Include="..\PubnubApi.Tests\EventEngine\HandshakingStateTransition.cs" Link="EventEngine\HandshakingStateTransition.cs" />
+    <Compile Include="..\PubnubApi.Tests\EventEngine\ReceiveFailedStateTransition.cs" Link="EventEngine\ReceiveFailedStateTransition.cs" />
+    <Compile Include="..\PubnubApi.Tests\EventEngine\ReceiveReconnectingStateTransition.cs" Link="EventEngine\ReceiveReconnectingStateTransition.cs" />
+    <Compile Include="..\PubnubApi.Tests\EventEngine\ReceiveStoppedStateTransition.cs" Link="EventEngine\ReceiveStoppedStateTransition.cs" />
+    <Compile Include="..\PubnubApi.Tests\EventEngine\ReceivingStateTransition.cs" Link="EventEngine\ReceivingStateTransition.cs" />
+    <Compile Include="..\PubnubApi.Tests\EventEngine\UnsubscribedStateTransition.cs" Link="EventEngine\UnsubscribedStateTransition.cs" />
     <Compile Include="..\PubnubApi.Tests\GenerateSampleGrant.cs" Link="GenerateSampleGrant.cs" />
     <Compile Include="..\PubnubApi.Tests\PubnubCommon.cs" Link="PubnubCommon.cs" />
     <Compile Include="..\PubnubApi.Tests\PubnubUnitTest.cs" Link="PubnubUnitTest.cs" />
@@ -110,6 +118,9 @@
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resource.Designer.cs</LastGenOutput>
     </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="EventEngine\" />
   </ItemGroup>
   <Target Name="PostBuild" AfterTargets="PostBuildEvent">
     <Exec Command="copy $(ProjectDir)fileupload.txt $(ProjectDir)$(OutDir)" />
